migrate blog page to have a better structure
