+++ .gitignore
@@ -19,3 +19,4 @@ stagewise.json
 *.tsbuildinfo
 /agents
 packages/dev-logger/tsconfig.tsbuildinfo
+*.tsbuildinfo
